Unicode-aware scanning of Unicode escapes
Unicode escape scanning needs to be beefed up somewhat to support correct
scanning of identifiers. I am talking here about word/symbol boundaries
which should be preserved when all Unicode characters are represented as
Unicode escapes.

The easy option for supporting Unicode escapes in identifiers would be
to allow literally anything in them (as it is now for chars and strings),
assume (in the scanner) that Unicode values are correct, just append them,
and go on scanning other characters. Errors will be reported later.
However, this approach has a flaw: Unicode escapes are not used to detect
identifier boundaries.

Suppose you have defined identifiers ℕ, ℤ, and ∖. Now you can write "ℤ∖ℕ"
in you program and it will be correctly scanned as "Word-Identifier(ℤ),
Symbol-Identifier(∖), Word-Identifier(ℕ)".

Now, someone cannot into Unicode, but they know about the ASCII fallback.
They simply escape any non-ASCII character in the source code and get
a string "\u{2124}\u{2216}\u{2115}". Which will be scanned incorrectly,
confuse the parser, and refuse to compile *if* the scanner does not see
the identifier boundaries here.

Thus, the scanner must be aware of the codepoints represented by Unicode
escapes. And given that it must parse them anyway, it can also validate
the range as well. And while we're here, we can also validate that \x??
escapes contain only ASCII values (represented as single bytes in UTF-8).

We return codepoint value only for Unicode escapes as they will be needed
by identifier scanning routines. Byte escapes can be used only in strings
and characters. They are not processed here, as C-style escapes are not
processed here, and newlines are not normalized here.