Scanning identifiers
All three kinds are implemented:

  - Words that are meant for general usage.

  - Marks that are meant for infix, postfix, and prefix operators.

  - Quotes that are meant for outfix operators.

With full Unicode support, Unicode escapes, and somewhat reasonable error
recovery. I guess the recovery may be bad at some cases.

The most complex part of it are Unicode escapes. They are the main source
of parsing errors.

The main deficiency is the lack of peeking into Unicode escapes. Peeking
is required for correct implementation of boundaries between identifier
kinds (so that "a+b" is scanned as three identifiers). Currently, there
is no way to peek an escape: we just scan over them, report and errors,
and then pretend that the escape value was in self.cur. This works bad
because we must 'unread' the escape if it break the identifier we are
currently scanning and start some new one.

Furthermore, complex syntax and boundary rules require multiple special
cases. Just look at the IdentifierSpecials enum to get an idea.

Also, Unicode escapes are prohibited for ASCII characters as this could
become a major sort of abuse. I decided that it is simpler to just make
them invalid rather than add even more complexity for handling invalid
code in some reasonable way. There is no way that a sane programmer will
ever want to use Unicode escapes for ASCII, so this should not matter in
practice.