Explicit reporting of invalid Unicode escapes
U+FFFD REPLACEMENT CHARACTER turned out to be a bad idea as it has
General_Category 'So' which is valid in symbol identifiers. Well,
I *thought* it had a different one...

Therefore, scan_unicode_escape() needs to explicitly state the fact
that the escape sequence is invalid. Option<T> to the rescue.

Though, invalid Unicode characters are still replaced with U+FFFD
in character and string literals. We need some character in there,
but we actually do not care which one. It will be important only
for identifiers.