Scan literal type suffixes
All literals can have a type suffix: an optional word identifier that
immediately follows the literal, without any whitespace or delimiters
between them. This patch adds scanning of such suffixes for integers.
floats, characters, strings, and raw strings.

Note that only word identifiers are allowed to be type suffixes.
Identifiers of other kinds and genuinely invalid character sequences
are not scanned over by scan_optional_type_suffix().

Also note the special for raw strings. This is necessary to correctly
handle cases like r""r"" which should be scanned as two raw strings,
not as a raw string with type suffix 'r' followed by an empty string.

Invalid Unicode characters and Unicode escapes for ASCII characters
are also handled in a slightly hacky way. We must scan them over, and
thus we will report any irregularities in Unicode escapes themselves
(like missing braces), but we do not report any 'invalid identifier
characters there', instead we unread() some invalid character which
will be properly reported when scanning the next token. Note that it
is okay in this case to lose specific information about the type of
invalid character (IncorrectUnicodeEscape vs. AsciiUnicodeEscape)
because this will be the first character of the next token--which
automatically means Token::Unrecognized where we do not report
individual characters as 'invalid in identifiers', only the whole
Unrecognized token gets reported as invalid character sequence.