Report unrecognized sequences immediately
Initially I thought this was a good idea to just keep scanning if the
first character of the would-be identifer is invalid. However, this has
some implications for type suffixes of literal constants...

Type suffixes are going to be defined as word identifiers immediately
following a literal, such as in 12i32 or "some"_utf16. However! What
should the lexer do if the first character is not a correct word starter?
Well, when scanning an identifier it can just eat it and go on: it knows
that it should output either an identifier if it meets the starter
eventually, or an invalid sequence if it does not. The point is that the
scanner does not care whether it is a word  identifier or a mark one.

But in case of _optional word suffixes_ we do care. Because we do want to
scan 2+2 as Integer(2), Mark(+), Integer(2). Thus we must immediately know
whether we are scanning a word or not. Allowing unrecognized sequences of
Unicode characters at the beginning prevents us from doing this. If the
sequence is not followed by a word starter the type suffix code will have
to do something with that sequence, either treat is as 'invalid' suffix,
or forget it, or backtrack. All options are not really good here.

Prohibiting invalid starting sequences solves this issue. If a literal
is not immediately followed by a valid word starter then it has no type
suffix. And whatever shit goes next, it will be scanned later by next().

P.S. Invalid characters and sequences are still allowed in the middle of
identifiers and at the end of them. I guess it is okay to permit and
report this as people are more likely to do a random mistake on the way
rather that writing the whole identifier incorrectly. Anyway, these
Token::Unrecognized are going to be fatal lexing errors so that the parser
will not even try to make sense out of a possibly malformed token stream.