Remove StringScanner::tok field
I have finally replaced this stateful scanning with return values.
It is more logical to do so, and it will simplify the code for scanning
type suffixes as it will allow to use scan_identifier_word() without
unwanted side effects.

I do not really remember why I have chosen the stateful approach.
Maybe to avoid that 'if' in scan_number() or something.

By the way, this also fixes some compiler warnings.