Fix scanning of identifiers starting with 'r'
Well, yeah. When we see an 'r' we need to decide whether it is a raw
string or an identifier. It is a raw string if 'r' is followed by zero
or more hashes '#' followed by a double quote '"'. Otherwise it is the
first letter of a word identifier.

Unfortunately, raw strings have context-sensitive grammar which can be
represented with the following indexed grammar:

    RawString   -> 'r' Spelling[]
    Spelling[h] -> '#' Spelling[#h] '#'
    Spelling[h] -> '"' Content[h] '"'
    Content[h]  -> NonQuote Content[h]
    Content[h]  -> '"' Hash[h] Continue[h]
    Content[h]  -> empty
    Hash[#h]    -> '#' Hash[h]
    Hash[#]     -> empty
    Continue[h] -> NonHash Content[h]
    Continue[h] -> empty
    NonQuote    -> {any character except '"'}
    NonHash     -> {any character except '#'}

Thus we need infinite lookahead to parse it and to disambiguate it from
an identifier 'r' followed by some hashes followed by something that is
not a double quote. Hashes are a part of multipart identifier syntax, so
we cannot simply eat them as a part of wannabe word-identifier starting
with an 'r'.