Scanning hash literals
These will be used in attribute syntax as well as in multipart
identifiers. I don't remember why I did not include them initially, but
here they are. It's just another punctuation character like ( or ;

The important thing is that it is treated as a terminator for identifiers
right now, becaue we will need this to correctly disambiguate between
a raw string and an identifier when we see a letter 'r'. Currently, the
scanner panics when it encounters an identifier starting with 'r'.