Scanning explicit symbols
This is an explicit form of symbols which allows arbitrary characters to
be included into them. It is similar to string delimited by backquotes,
but has some differences. Notably, symbols cannot contain explicit line
endings or line ending escapes. They also have their own special escape
sequence for the backquote character. Finally, symbols are not intended
to be 'immutable strings', thus they are not literal values and do not
have any type suffixes. Though, they *are* immutable and have an inherent
string value.

The scanner implementation is similar to those of characters and strings:
it's the same copy-pasta of scan-one-character and scan-escape-sequence
code. Maybe we should do something about it...

I have also had to fixup identifier scanning code to treat backquote (`)
as a terminator character. The tests have also been updated accordingly.