Scanning documentation comments
These are special kind of comments which can be automatically processed by
the compiler to attach documentation to program entities. Their syntax is
similar to what Doxygen and Rustdoc understand:

    /// Line comment for the following entity

    //! Line comment for the enclosing entity

    /**
     *  Block comment for the following entity.
     */

    /*!
     *  Block comment for the enclosing entity.
     */

However, akin to Rustdoc, Sash does not consider these things
to be documentation comments:

    /////////////////////////////////////////////
    /********************************************
    **                                         **
    ** I love to write heavy-to-maintain boxes **
    **                                         **
    ********************************************/
    /////////////////////////////////////////////

A doc comment starts with exactly '///' followed by a non-'/' which is
the preferred form. Block comment is an alternative form which must start
with '/**' followed by a non-'*'. Enclosing documentation comments may use
whatever they want after '!' as this exclamation mark already tells us
that this is not a separator line.