Update project layout
I have split the current code into two separate libraries: libsyntax and
libunicode. Technically, I guess only libsyntax will actually use
libunicode so it could be a simple module inside the libsyntax.
However, I was not able to compile and run the documentation example from
it.

For some reason, rustdoc wants to run that example because sash_identifiers
is public. However, the 'unicode' module itself is not, and the doctest
runner is not in the 'unicode' module so it has not access to the methods
of sash_identifier. It seems the core issue is that Rust does not have
crate-level visibility specified (similar to package visibility in Java).

I do not want to make 'unicode' public in libsyntax so I have looked
around in the internets and the only solution I found was to use multiple
crates. So that's what I did.

Unfortunately, cargo does not have a way to recusively run tests for a
crate and its dependencies, so the current test runner will do additional
crate recompilation. Meh. This should not be an issue in practice.