Relax lifetime restrictions on SpanReporter
SpanReporter should not have been tied to the lifetime of error messages
in the first place. Currently, error messages are all 'static, but they
may be (and will be) generated by the lexer. We do not need to keep the
lexer around to use the tokens or error messages returned by it.

If a particular implementation of SpanReporter needs to keep the string
around for some time then it should just copy it and do whatever it wants.

Another part of this commit removes error messages from tests. I gave this
a thought and decided to not test the error content automatically. This
would either require first-classing error kinds into an enum, or declaring
a bunch of weird 'const &str' for the messages.

First-classing is not useful anywhere outside of the tests. The parser and
the error reporting machinery are the main users of the lexer. And they
do not really care about the nature of a particular error as they do not
make decisions based on it. The parser just needs to know whether an error
is fatal for it, and the reporter needs to know the priority of an error
to decide how it should print out the message and whether it needs to be
printed out at all. Both of them don't give a shit was it a missing brace,
an invalid Unicode, or a bare CR in a raw string.

Using constants for messages looks like a way to go, but the problem is
that checking error messages does not add any real value to tests. There
are not many errors which can arise in exactly the same span of text, so
there is not much incentive to test which one of them we caught. It is
enough to validate the count and locations of the errors.