Upgrade the only warning to error
There is actually no reason to produce warnings from scanner because
warnings are used for potentially incorrect code. *Proper* syntactical
analysis never produces potentially incorrect parses. Either the code
is syntactically valid, or it is a syntax error.

The errors may be fatal or recoverable, but that's a different story.
This will be handled by further commits.