Upgrade some errors to Fatal
The rule of thumb: if it produces Token::Unrecognized then it is fatal.
This includes various 'missing terminator' errors which are detected
only after the scanner may have consumed some excess characters from
the tokens which follow the erroneous one.

There is one exception to this rule: the pokemon error produced by the
identifier scanning code ("Incomprehensible character sequence") which
appears when the source code contains a Unicode sequence that does not
start with something that can start a valid token. This sequence is
scanned over until the first comprehensible character, and thus it is
kinda like an identifier (and is likely to be a malformed one). So we
do have a possibility to recover from it by treating the Unrecognized
token as an Identifier if we expect one (or by skipping it otherwise).