String intern pool
Well, the module description says it all. This will be needed for
efficient storage and comparison of lots of strings which we will be
dealing with.

The implementation is quite straightforward: HashMap is used for fast
checks string presence in the pool and for fast lookup of associated
atom values, Vec is used for fast lookup of string values by atoms,
the immutable string values are shared between these two containers
and can be leased to outside clients.

The pool is not synchornized whatsoever so it cannot be used concurrently
and Atoms are !Send. Though, I do not intend to have a thread-local global
intern pool as rustc does. Rather, a pool should be kept around the atoms
it produces, and must be discarded (freeing up the memory used by strings)
as soon as it is not needed.