Remove start/end fields from StringScanner
There is actually no need in them. Subroutines that need this information
can track it themselves using their own stack frames.

The only tricky part here are raw strings. We do not have direct
information about the starting index of a raw string literal, but we do
have coordinates of the double quote character, number of hashes in this
literal, and we know that 'r' and '#' are ASCII characters which take
exactly 1 byte in UTF-8. So we can simply calculate the starting point.