Move generated code into tables.rs
We're going to have more generated code and--more importantly--more
hand-written code. Thus it will not be practical to keep the whole
implementation in a single file.

I expect to have tables.rs contain *all* automatically generated code
which handles Unicode character data tables. All other modules will be
made for human beings.

The current interface is preserved: the identifier definitions are simply
reexported from tables::sash_identifiers. I would like to split the code
and tables apart, but the tables are very dependent on the actual
definition of the Category type, thus it is not practical to have the
apart.