Normalization conformance test and benchmarks
This big ass patch adds tests (based on NormalizationTest.txt from
UCD) to confirm quality of the implementation, and some benchmarks
to measure its performance. The tests are automatically generated
from UCD data and take a lot of space. The benchmarks are
handwritten though.

Both files use all-escaped strings to avoid silly issues with editors
that might think that normalizing the whole source code is a good
thing to do. We _are_ testing normalization here and we _do_ want
denormalized strings exactly as they are written. So the files are
kept ASCII-only.

Also, they are put into separate 'tests' and 'benches' directories,
so Cargo will treat them as integration tests. The main issue here
is that normalization_conformance.rs takes huge amount of time to
compile, and normalization_benchmark.rs (obviously) uses unstable
#![feature(test)] so it must not break stable builds.