Initial normalization implementation
This is pretty straightforward (and slow) implementation of the
normalization algorithms. However, it works. And even does not
fail the conformance test!

We'll be working on performance improvements in some next commits.
The main bottlenecks that need to be taken care of:

  - canonical combining classes are recalculated all the time

  - binary search is reasonably fast and compact, but there are
    other ways to organize data in order to trade memory for speed

  - we'll be dealing with lots of prenormalized data so quick check
    is a must

All these things should make the implementation fast enough for us
to be not bothered with performance for a long time. Also, I would
care for making the tables to be more compact.