Cache character canonical combining classes
First of all, there too many calls to ccc() in the code. Each CCC lookup
means a binary search over a 6 KB table (read: it is slow). We repeatedly
recalculate CCCs for characters, but actually we can do this only once.
Furthermore, we can precalculate CCCs for compositions and decompositions
and save some more time on this. So we replace each char with (char, u8)
and keep the CCC around, filtering it when we allocate a String object.

Minor bonus: we also don't have to do lookups for Hangul syllables as they
have been all assigned CCC zero which will never change.