Use trie for decomposition lookup
We use the same approach (and code) as for the canonical combining
classes to lookup decompositions. Decomposition tables are used by
all normalization forms so this impacts everything. Furthermore,
they are quite large (read: binary search needs quite a few probes,
i.e. cache misses). Converting them to a trie yields ~25% increase
in average performance of the normalization benchmark on my machine.

A single trie is used for both canonical and compatibility mappings
in order to conserve space. The trie uses 321 KB for tables. Binary
search used up 306 KB (but it was somewhat inefficient and could be
squeezed to 256-something KB by gathering character data in one
giant slice).

We also do not use hash table again as it would take too much space
in order to be efficient. Unicode 8.0.0 has 11773 distinct mappings,
each needs a 12-byte bucket (character:4, 2 x (offset:2, length:1,
padding: 1)), and the hash table needs to have at most 30% load to
provide efficient lookups (at least as fast as the trie). This gives
us 0.75 MB for a hash table with 65536 buckets, plus 192 KB more for
the actual character data, summing to 960 KB for decomposition tables.
Increase this to 1.70 MB if you want to be faster that the trie.