Use hash table for primary composite lookup
Finally! Unleash the power of the greatest invention of computer science
since the sliced bread! (Yeah, I really wanted to get a hash table.)

This replaces binary search with a hash table for looking up composites.
Somewhere where it is really approriate. The table is not that big, and
it is definitely smaller than a double-layered trie. The binary search
tables need 14.7 KB. The hash table needs 64 KB. A trie will need...
who cares? I just believe that it is smaller and am fine with that.

The table uses a simple multiplicative hash function with precomputed
coefficients. This is made for consistency of generated results. The
code includes an option to generate and analyze new hash functions.

The coefficients seems to work quite well for Unicode 8.0.0, giving 25%
miss rate on average, and maximum of 3 lookups needed to determine that
a code point pair is not a primary composite (which is the most common
case). The table uses increments for collision resolution, so that all
buckets for a single pair are adjacent and fit into a single cache line.

In the end, this yields ~23% speedup on my machine in comparison with
the previous binary search approach.