Compactify 'char + CCC' tuple layout
This introduces a helper type `charcc` which is an optimized layout of
`(char, u8)` tuples used to represent chars with precomputed CCC values.
As explained in the docstring, this reduces the size of 'char + CCC'
from 8 bytes to just 4. This reduces the size of data tables by 110 KB.

The new type also shuffles assertions a bit, reducing the number or
checks made in Hangul processing, and increasing it in other parts.
This leads to 15% overall benchmark speedup consisting of 31% speedup
of Hangul processing and 5% slowdown of everything else. We are going
to take care of this in the next patch.