Use Quick_Check to speed up normalization
I have considered implementing the quick check as either simple
binary search over excluded ranges, or using a trie to filter out
excluded characters. There are quite a few consecutive ranges in
these tables, so binary search turns out to be very compact.
However, it slows down processing in some cases, so I ended up
choosing the trie approach.

Benchmarking this is pretty dumb as the effectiveness of quick
check is very depedent on the script of the text. If combining
modifiers are normal in the script then quick check will do
nothing for C and KC forms as most of the characters have MAYBE
value in their Quick_Check.

In any case, the main booster here is the ASCII check,
everything else is just a bonus.

Here are some averages:

  Speed (relative to baseline, %):      Binary size bloat (KB):

                 bsearch    trie          bsearch   trie
  ------------------------------          --------------
  normalized      +37.69  +58.93          +8.27   +19.65
  denormalized     -9.15   +0.39
  altogether      +12.40  +26.13