Remove ZWNJ and ZWJ during normalization
I planned doing this since the beginning as a part of normalization,
but I forgot. Now zero-width joiner and zero-width non-joiner code
points are removed from normalized forms of identifiers.

As usual, they remain intact in strings and explicit symbols, and
are of course allowed in character literals, but not in word
identifiers or implicit symbols. Well, they are allowed there, but
they get stripped during normalization.

ZWNJ/ZWJ usage is understandable in Arabic or Hindu texts, but
it is a source of security risks in other scripts. While UAX #31
has provisions for this (in the form of script restriction), and
UTR #36 has much more to say about security, I do not plan to
complicate Unicode support much more to handle *everything*.

So we will simply remove ZWNJ and ZWJ from normalized forms. The
tests have also been adjusted to this, and the table generator
already does all it needs to.