Basic keyword support in the scanner
The parser will need to handle keywords and hardcoding atoms is not
good, so we introduce a new token type to represent keywords which is
derived from identifiers. The keywords are parsed *after* identifier
normalization to disallow keyword usage in any form and preserve the
closure property over NFKC application to all identifiers.

We introduce two keywords for now: 'module' and 'library'. These will
be used in the initial parser stub.