Introduce tree diffs
These will be used for parser output verifications. The flat diff
algorithm should work reasonably well.

Note that the comparator comes as an argument to the differ. This
allows mulitple implementations of a comparator. TreeNode trait,
on the other hand, can have only one implementation for a single
type. This is intentional as the set of child nodes is usually an
intrinsic property of a tree node (i.e., there is only one way to
get a set of child nodes).

TreeNode trait is placed into a separate module as it is useful
not only for diffing. It's just a generic trait of tree-like data.