Formatting tree diffs
Initially I wanted to do something with fancier alignment,
like this:

  A
  |- B
  |  |-     C
  |  |- (+) D
  |  `-     E
  |- (+) F
  `-     G

But I could not formalize my desires, and this would require non-
context-free formatting of tree nodes, so the generic formatter
would not work. So we end up with more boring and harder to read
variant:

  A
  |- B
  |  |- C
  |  |- (+) D
  |  `- E
  |- (+) F
  `- G

Also, we have a pruning function to remove non-interesting all-equal
branches of the diff tree, which leaves only essential information.